requirements.txt
aiohttp==3.6.2
aiohttp-sse-client==0.1.7
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
crc16==0.1.1
cryptography==2.9.2
Django==2.2.14
django-cors-headers==3.4.0
django-environ==0.4.5
django-model-utils==4.0.0
django-polaris==0.12.0
djangorestframework==3.11.0
idna==2.10
mnemonic==0.19
multidict==4.7.6
psycopg2-binary==2.8.5
pycparser==2.20
PyJWT==1.7.1
PyNaCl==1.3.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
stellar-base-sseclient==0.0.21
stellar-sdk==2.6.1
toml==0.10.1
urllib3==1.25.9
whitenoise==5.1.0
yarl==1.4.2
